refactor: 페키지 이름 변경
