--- backend/src/test/java/wooteco/prolog/login/ui/auto_interceptor_uri_mapper/test_classes/NormalClass.java
@@ -1,4 +1,0 @@
-package wooteco.prolog.login.ui.auto_interceptor_uri_mapper.test_classes;
-
-public class NormalClass {
-}
