+++ backend/src/test/java/wooteco/prolog/login/ui/autoceptor/ClassScannerTest.java
@@ -1,11 +1,11 @@
-package wooteco.prolog.login.ui.auto_interceptor_uri_mapper;
+package wooteco.prolog.login.ui.autoceptor;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import wooteco.prolog.login.ui.auto_interceptor_uri_mapper.scanner.ClassScanner;
-import wooteco.prolog.login.ui.auto_interceptor_uri_mapper.test_classes.ControllerClass;
-import wooteco.prolog.login.ui.auto_interceptor_uri_mapper.test_classes.RestControllerClass;
-import wooteco.prolog.login.ui.auto_interceptor_uri_mapper.test_classes.NormalClass;
+import wooteco.prolog.login.ui.autoceptor.scanner.ClassScanner;
+import wooteco.prolog.login.ui.autoceptor.test_classes.ControllerClass;
+import wooteco.prolog.login.ui.autoceptor.test_classes.RestControllerClass;
+import wooteco.prolog.login.ui.autoceptor.test_classes.NormalClass;
 
 import java.util.Set;
 
@@ -17,7 +17,7 @@ class ClassScannerTest {
     @Test
     void getAllClasses() {
         // given
-        final String basePackage = "wooteco.prolog.login.ui.auto_interceptor_register.test_classes";
+        final String basePackage = "wooteco.prolog.login.ui.autoceptor.test_classes";
         ClassScanner classScanner = new ClassScanner(basePackage);
 
         // when
