+++ backend/src/test/java/wooteco/prolog/login/ui/autoceptor/test_classes/ControllerClass.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.login.ui.auto_interceptor_uri_mapper.test_classes;
+package wooteco.prolog.login.ui.autoceptor.test_classes;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
