+++ backend/src/test/java/wooteco/prolog/login/ui/autoceptor/test_classes/NormalClass.java
@@ -1,0 +1,4 @@
+package wooteco.prolog.login.ui.autoceptor.test_classes;
+
+public class NormalClass {
+}
